+++ packages/rocketh-verifier/src/metadata.ts
@@ -2,20 +2,20 @@ import {loadDeployments, ResolvedUserConfig} from 'rocketh';
 import fs from 'fs-extra';
 import path from 'path';
 
-export function exportMetadata(config: ResolvedUserConfig, targetName: string, {out}: {out: string}) {
-	const {deployments, chainId} = loadDeployments(config.deployments, targetName, false);
+export function exportMetadata(config: ResolvedUserConfig, environmentName: string, {out}: {out: string}) {
+	const {deployments, chainId} = loadDeployments(config.deployments, environmentName, false);
 
 	if (Object.keys(deployments).length === 0) {
-		console.log(`the target ${targetName} has zero deployments`);
+		console.log(`the environment ${environmentName} has zero deployments`);
 		process.exit();
 	}
 
 	if (!chainId) {
-		console.error(`the target ${targetName} has no chainId recorded`);
+		console.error(`the environment ${environmentName} has no chainId recorded`);
 		process.exit(1);
 	}
 
-	const folder = path.join(out, targetName);
+	const folder = path.join(out, environmentName);
 	fs.emptyDirSync(folder);
 	const deploymentNames = Object.keys(deployments);
 	for (const deploymentName of deploymentNames) {