+++ packages/rocketh-viem/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @rocketh/viem
 
+## 0.15.0-testing.2
+
+### Patch Changes
+
+- rname target to environment
+
 ## 0.15.0-testing.1
 
 ### Patch Changes