+++ packages/rocketh/CHANGELOG.md
@@ -1,5 +1,11 @@
 # rocketh
 
+## 0.15.0-testing.13
+
+### Patch Changes
+
+- rname target to environment
+
 ## 0.15.0-testing.12
 
 ### Patch Changes